🔧 회원 가입 시 작가 승인 부분 수정 및 파일 위치 옮기기
