+++ backend/src/main/java/com/ome/dto/auth/request/SignupRequestDto.java
@@ -12,7 +12,6 @@ public class SignupRequestDto {
 	private String username;
 	private String password;
 	private String email;
-	private Role role;
-	
+	private boolean applyAsCreator;
 
 }
