+++ backend/src/main/java/com/ome/dto/mypage/response/AdminMyPageResponseDto.java
@@ -1,4 +1,4 @@
-package com.ome.dto.auth.response;
+package com.ome.dto.mypage.response;
 
 
 
@@ -7,22 +7,21 @@ import com.ome.domain.Users;
 import lombok.Getter;
 
 //🌟🌟 UserService -> 권한에 따른 마이페이지 분기 
-// 작가 마이페이지 dto !!!
+//관리자 마이페이지 dto !!!
 @Getter
-public class CreatorMyPageResponseDto {
-	
+public class AdminMyPageResponseDto {
 	private String userId;
 	private String username;
 	private String email;
-	private int subscriverCount;
-	private int recipeCount;
+	private int totlaUsers;
+	private int pendingCreators;
 	
-	public CreatorMyPageResponseDto (Users user) {
+	public AdminMyPageResponseDto(Users user) {
 		this.userId = user.getUserId();
 		this.username = user.getUsername();
 		this.email = user.getEmail();
-		this.subscriverCount = 0;
-		this.recipeCount = 0;
+		this.totlaUsers = 0;
+		this.pendingCreators = 0;
 	}
 
 }
