+++ backend/src/main/java/com/ome/service/auth/AuthService.java
@@ -1,9 +1,11 @@
 package com.ome.service.auth;
 
 import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.security.web.server.header.CacheControlServerHttpHeadersWriter;
 import org.springframework.stereotype.Service;
 
 import com.ome.domain.Users;
+import com.ome.domain.Users.Role;
 import com.ome.dto.auth.request.LoginRequestDto;
 import com.ome.dto.auth.request.SignupRequestDto;
 import com.ome.dto.auth.response.SignupResponseDto;
@@ -35,19 +37,28 @@ public class AuthService {
 			throw new IllegalArgumentException("이미 존재하는 아이디입니다.");
 		}
 		
+		Role role = Role.USER;
+		boolean approved = true;
+
+		// 작가 신청 시 -> 임시로 user로 표시하고 관리자 승인 대기함.
+		if (dto.isApplyAsCreator()) {
+			role = Role.USER;       
+			approved = false;        
+		}
+		
 		Users user = Users.builder()
 				.userId(dto.getUserId())
 				.username(dto.getUsername())
 				.email(dto.getEmail())
 				.password(passwordEncoder.encode(dto.getPassword())) // 비밀번호 암호화
-				.role(dto.getRole()) // ROLE_USER와 ROLE_CREATOR만 허용
-				.approved(false) // 작가 승인 default 값으로 false 지정
+				.role(role) // ROLE_USER와 ROLE_CREATOR만 허용
+				.approved(approved) // 작가 승인 default 값으로 false 지정
 				.build();
 		
 		repository.save(user);	
 		
 		//토큰 생성하기 
-		String token = jwtUtil.createToken(dto.getUserId(),dto.getRole().name());
+		String token = jwtUtil.createToken(user.getUserId(),user.getRole().name());
 		return new SignupResponseDto("회원가입 성공", token);
 		
 	}
